+++ kipper/core/src/compiler/semantics/const.ts
@@ -139,6 +139,20 @@ export type KipperListType<ValueType extends KipperType> = "list";
 export const kipperListType: KipperListType<any> = "list";
 
 /**
+ * Represents the meta type in Kipper, which itself is used represents a type e.g. this is the type of a
+ * type.
+ * @since 0.8.0
+ */
+export type KipperMetaType = "type";
+
+/**
+ * Represents the meta type in Kipper, which itself is used represents a type e.g. this is the type of a
+ * type.
+ * @since 0.8.0
+ */
+export const kipperMetaType: KipperMetaType = "type";
+
+/**
  * String-like types that include both char and string.
  * @since 0.6.0
  */
@@ -148,7 +162,7 @@ export type KipperStrLikeTypes = KipperStrType | KipperCharType;
  * String-like types that include both char and string.
  * @since 0.6.0
  */
-export const kipperStrLikeTypes: Array<KipperStrType | KipperCharType> = [kipperStrType, kipperCharType];
+export const kipperStrLikeTypes: Array<KipperStrLikeTypes> = [kipperStrType, kipperCharType];
 
 /**
  * All primitive types inside Kipper.
@@ -165,12 +179,12 @@ export const kipperPrimitiveTypes = [kipperVoidType, kipperNumType, kipperStrTyp
 /**
  * All available variable types inside Kipper.
  */
-export type KipperType = KipperFuncType | KipperPrimitiveType | KipperListType<any>;
+export type KipperType = KipperMetaType | KipperFuncType | KipperPrimitiveType | KipperListType<any>;
 
 /**
  * All available variable types inside Kipper.
  */
-export const kipperTypes: Array<string> = [kipperFuncType, ...kipperPrimitiveTypes, kipperListType];
+export const kipperTypes: Array<string> = [kipperMetaType, kipperFuncType, ...kipperPrimitiveTypes, kipperListType];
 
 /**
  * Types that may be returned by a function.
@@ -183,6 +197,20 @@ export type KipperReturnType = KipperPrimitiveType | KipperListType<any>;
  * @since 0.6.0
  */
 export const kipperReturnTypes: Array<string> = [...kipperPrimitiveTypes, "list"];
+
+/**
+ * List of all supported variable type conversions that can be performed in a Kipper program.
+ *
+ * For each translation, there will have to be a corresponding {@link KipperTargetBuiltInGenerator generator function},
+ * which generates for each conversion the translator function in the specific target.
+ * @since 0.8.0
+ */
+export const kipperSupportedConversions: Array<[KipperType, KipperType]> = [
+	["num", "str"],
+	["str", "num"],
+	["bool", "str"],
+	["bool", "num"],
+];
 
 /**
  * All available storage types inside Kipper.