+++ kipper/core/src/compiler/semantics/tokens/statements.ts
@@ -25,7 +25,7 @@ import type { KipperScope, TranslatedCodeLine } from "../const";
 import type { VariableDeclaration } from "./definitions";
 import type { Expression } from "./expressions";
 import type { TargetTokenCodeGenerator } from "../../translation";
-import type { TargetTokenSemanticAnalyser } from "../semantic-analyser";
+import type { TargetTokenSemanticAnalyser } from "../target-semantic-analyser";
 
 /**
  * Every antlr4 statement ctx type
@@ -191,7 +191,7 @@ export class CompoundStatement extends Statement<{ scope: KipperScope }> {
 	}
 
 	/**
-	 * Performs type checking for this Kipper token.This will log all warnings using {@link programCtx.logger}
+	 * Performs type checking for this Kipper token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 * @since 0.7.0
 	 */
@@ -249,7 +249,7 @@ export class SelectionStatement extends Statement<{ scope: KipperScope }> {
 	}
 
 	/**
-	 * Performs type checking for this Kipper token.This will log all warnings using {@link programCtx.logger}
+	 * Performs type checking for this Kipper token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 * @since 0.7.0
 	 */
@@ -306,7 +306,7 @@ export class ExpressionStatement extends Statement<{ scope: KipperScope }> {
 	}
 
 	/**
-	 * Performs type checking for this Kipper token.This will log all warnings using {@link programCtx.logger}
+	 * Performs type checking for this Kipper token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 * @since 0.7.0
 	 */
@@ -364,7 +364,7 @@ export class IterationStatement extends Statement<{ scope: KipperScope }> {
 	}
 
 	/**
-	 * Performs type checking for this Kipper token.This will log all warnings using {@link programCtx.logger}
+	 * Performs type checking for this Kipper token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 * @since 0.7.0
 	 */
@@ -422,7 +422,7 @@ export class JumpStatement extends Statement<{ scope: KipperScope }> {
 	}
 
 	/**
-	 * Performs type checking for this Kipper token.This will log all warnings using {@link programCtx.logger}
+	 * Performs type checking for this Kipper token. This will log all warnings using {@link programCtx.logger}
 	 * and throw errors if encountered.
 	 * @since 0.7.0
 	 */