+++ kipper/core/src/targets/typescript/code-generator.ts
@@ -18,12 +18,14 @@ import {
 	FStringPrimaryExpression,
 	FunctionCallPostfixExpression,
 	FunctionDeclaration,
+	GenericTypeSpecifierExpression,
 	IdentifierPrimaryExpression,
 	IncrementOrDecrementExpression,
 	IncrementOrDecrementUnaryExpression,
 	IterationStatement,
 	JumpStatement,
 	KipperTargetCodeGenerator,
+	KipperType,
 	ListPrimaryExpression,
 	LogicalAndExpression,
 	LogicalOrExpression,
@@ -34,14 +36,18 @@ import {
 	RelationalExpression,
 	ScopeFunctionDeclaration,
 	SelectionStatement,
+	SingleTypeSpecifierExpression,
 	StringPrimaryExpression,
 	TangledPrimaryExpression,
+	TargetTokenCodeGenerator,
 	TranslatedCodeLine,
 	TranslatedCodeToken,
 	TranslatedExpression,
+	TypeofTypeSpecifierExpression,
 	VariableDeclaration,
 } from "../../compiler";
 import { getTypeScriptBuiltInIdentifier, getTypeScriptType } from "./tools";
+import { getConversionFunctionIdentifier } from "../../utils";
 
 /**
  * The TypeScript target-specific code generator for translating Kipper into TypeScript.
@@ -165,6 +171,27 @@ export class TypeScriptTargetCodeGenerator extends KipperTargetCodeGenerator {
 	};
 
 	/**
+	 * Translates a {@link SingleTypeSpecifierExpression} into the typescript language.
+	 */
+	singleTypeSpecifierExpression = async (token: SingleTypeSpecifierExpression): Promise<TranslatedExpression> => {
+		return [];
+	};
+
+	/**
+	 * Translates a {@link GenericTypeSpecifierExpression} into the typescript language.
+	 */
+	genericTypeSpecifierExpression = async (token: GenericTypeSpecifierExpression): Promise<TranslatedExpression> => {
+		return [];
+	};
+
+	/**
+	 * Translates a {@link TypeofTypeSpecifierExpression} into the typescript language.
+	 */
+	typeofTypeSpecifierExpression = async (token: TypeofTypeSpecifierExpression): Promise<TranslatedExpression> => {
+		return [];
+	};
+
+	/**
 	 * Translates a {@link StringPrimaryExpression} into the typescript language.
 	 */
 	stringPrimaryExpression = async (token: StringPrimaryExpression): Promise<TranslatedExpression> => {
@@ -221,7 +248,7 @@ export class TypeScriptTargetCodeGenerator extends KipperTargetCodeGenerator {
 		const semanticData = token.ensureSemanticDataExists();
 		const func = token.programCtx.semanticCheck(token).getExistingFunction(semanticData.identifier);
 
-		// Add lib identifier prefix '_kipperGlobal_'
+		// Get the proper identifier for the function
 		const identifier =
 			func instanceof ScopeFunctionDeclaration ? func.identifier : getTypeScriptBuiltInIdentifier(func.identifier);
 
@@ -257,7 +284,20 @@ export class TypeScriptTargetCodeGenerator extends KipperTargetCodeGenerator {
 	 * Translates a {@link CastOrConvertExpression} into the typescript language.
 	 */
 	castOrConvertExpression = async (token: CastOrConvertExpression): Promise<TranslatedExpression> => {
-		return [];
+		// Get the semantic data
+		const semanticData = token.ensureSemanticDataExists();
+
+		const exp: TranslatedExpression = await semanticData.exp.translateCtxAndChildren();
+		const originalType: KipperType = semanticData.exp.ensureSemanticDataExists().evaluatedType;
+		const destType: KipperType = semanticData.type;
+
+		if (originalType === destType) {
+			// If both types are the same we will only return the translated expression to avoid useless conversions.
+			return exp;
+		} else {
+			const func: string = getTypeScriptBuiltInIdentifier(getConversionFunctionIdentifier(originalType, destType));
+			return [func, "(", ...exp, ")"];
+		}
 	};
 
 	/**