+++ test/module/core/compiler.test.ts
@@ -10,6 +10,7 @@ import {
 import { promises as fs } from "fs";
 import * as ts from "typescript";
 import * as path from "path";
+import { getTypeScriptBuiltInIdentifier } from "@kipper/core/lib/targets/typescript/tools";
 
 // Test files
 const mainFile = path.resolve(`${__dirname}/../../kipper-files/main.kip`);
@@ -25,6 +26,7 @@ const assignFile = path.resolve(`${__dirname}/../../kipper-files/assign.kip`);
 const addedHelloWorldFile = path.resolve(`${__dirname}/../../kipper-files/added-hello-world.kip`);
 const assignmentArithmeticsFile = path.resolve(`${__dirname}/../../kipper-files/assignment-arithmetics.kip`);
 const boolFile = path.resolve(`${__dirname}/../../kipper-files/bool.kip`);
+const typeConversion = path.resolve(`${__dirname}/../../kipper-files/type-conversion.kip`);
 
 describe("KipperCompiler", () => {
 	describe("constructor", () => {
@@ -130,6 +132,12 @@ describe("KipperCompiler", () => {
 
 			it("Variable assignment", async () => {
 				const fileContent = (await fs.readFile(assignmentArithmeticsFile, "utf8" as BufferEncoding)).toString();
+				const stream = new KipperParseStream(fileContent);
+				await compiler.syntaxAnalyse(stream);
+			});
+
+			it("Type conversion", async () => {
+				const fileContent = (await fs.readFile(typeConversion, "utf8" as BufferEncoding)).toString();
 				const stream = new KipperParseStream(fileContent);
 				await compiler.syntaxAnalyse(stream);
 			});
@@ -337,6 +345,22 @@ describe("KipperCompiler", () => {
 
 				assert(instance.programCtx);
 				assert(instance.programCtx.stream === stream, "Expected matching streams");
+			});
+
+			it("Type conversion", async () => {
+				const fileContent = (await fs.readFile(typeConversion, "utf8" as BufferEncoding)).toString();
+				const stream = new KipperParseStream(fileContent);
+				const instance: KipperCompileResult = await compiler.compile(stream);
+
+				assert(instance.programCtx);
+				assert(instance.programCtx.internals);
+
+				const code = instance.write();
+				assert(code);
+				assert(code.includes(getTypeScriptBuiltInIdentifier("strToNum")));
+				assert(code.includes(getTypeScriptBuiltInIdentifier("numToStr")));
+				assert(code.includes(getTypeScriptBuiltInIdentifier("boolToStr")));
+				assert(code.includes(getTypeScriptBuiltInIdentifier("boolToNum")));
 			});
 		});